Constraints test hierarchies replaced with JUnit Suite runner
